requirements.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
flake8==2.0
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
gevent-zeromq==0.2.5
greenlet==0.4.1
mccabe==0.2.1
pep8==1.4.6
psycopg2==2.5.1
pyflakes==0.7.3
pyzmq==2.2.0
wsgiref==0.1.2
